Add more tests for COMPRESS package
