+++ unittest/formats/test_dump_cfg.cpp
@@ -21,6 +21,8 @@
 
 using ::testing::Eq;
 
+bool verbose = false;
+
 class DumpCfgTest : public MeltTest {
     std::string dump_style = "cfg";
